refactor to create SmithyHTTPBindingProtocolGenerator
