+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/middlewares/OperationEndpointResolverMiddleware.kt
@@ -212,7 +212,7 @@ open class OperationEndpointResolverMiddleware(
         }
     }
 
-    open fun getBuiltInName(param: Parameter): String {
+    fun getBuiltInName(param: Parameter): String {
         return param.builtIn.get().split("::").last().toLowerCamelCase()
     }
