+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/protocols/rpcv2cbor/RpcV2CborProtocolGenerator.kt
@@ -6,61 +6,30 @@ import software.amazon.smithy.model.shapes.ShapeId
 import software.amazon.smithy.model.traits.StreamingTrait
 import software.amazon.smithy.model.traits.UnitTypeTrait
 import software.amazon.smithy.protocol.traits.Rpcv2CborTrait
-import software.amazon.smithy.rulesengine.language.EndpointRuleSet
-import software.amazon.smithy.rulesengine.traits.EndpointRuleSetTrait
-import software.amazon.smithy.rulesengine.traits.EndpointTestsTrait
-import software.amazon.smithy.swift.codegen.EndpointTestGenerator
 import software.amazon.smithy.swift.codegen.SyntheticClone
 import software.amazon.smithy.swift.codegen.integration.DefaultHTTPProtocolCustomizations
-import software.amazon.smithy.swift.codegen.integration.HTTPBindingProtocolGenerator
 import software.amazon.smithy.swift.codegen.integration.HttpBindingResolver
-import software.amazon.smithy.swift.codegen.integration.HttpProtocolClientGeneratorFactory
-import software.amazon.smithy.swift.codegen.integration.HttpProtocolTestGenerator
-import software.amazon.smithy.swift.codegen.integration.HttpProtocolUnitTestErrorGenerator
-import software.amazon.smithy.swift.codegen.integration.HttpProtocolUnitTestRequestGenerator
-import software.amazon.smithy.swift.codegen.integration.HttpProtocolUnitTestResponseGenerator
 import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
-import software.amazon.smithy.swift.codegen.integration.SmithyHttpProtocolClientGeneratorFactory
 import software.amazon.smithy.swift.codegen.integration.middlewares.ContentLengthMiddleware
 import software.amazon.smithy.swift.codegen.integration.middlewares.ContentTypeMiddleware
 import software.amazon.smithy.swift.codegen.integration.middlewares.MutateHeadersMiddleware
-import software.amazon.smithy.swift.codegen.integration.middlewares.OperationEndpointResolverMiddleware
 import software.amazon.smithy.swift.codegen.integration.middlewares.OperationInputBodyMiddleware
 import software.amazon.smithy.swift.codegen.middleware.MiddlewareRenderable
 import software.amazon.smithy.swift.codegen.model.getTrait
 import software.amazon.smithy.swift.codegen.model.hasTrait
 import software.amazon.smithy.swift.codegen.model.targetOrSelf
-import software.amazon.smithy.swift.codegen.testModuleName
 
 class RpcV2CborProtocolGenerator(
-    rpcCborCustomizations: DefaultHTTPProtocolCustomizations = RpcV2CborCustomizations(),
-    private val operationEndpointResolverMiddlewareFactory: ((ProtocolGenerator.GenerationContext, Symbol) -> MiddlewareRenderable)? = null,
-    private val userAgentMiddlewareFactory: ((ProtocolGenerator.GenerationContext) -> MiddlewareRenderable)? = null
-) : HTTPBindingProtocolGenerator(rpcCborCustomizations) {
+    customizations: DefaultHTTPProtocolCustomizations = RpcV2CborCustomizations(),
+    operationEndpointResolverMiddlewareFactory: ((ProtocolGenerator.GenerationContext, Symbol) -> MiddlewareRenderable)? = null,
+    userAgentMiddlewareFactory: ((ProtocolGenerator.GenerationContext) -> MiddlewareRenderable)? = null
+) : SmithyHTTPBindingProtocolGenerator(
+    customizations,
+    operationEndpointResolverMiddlewareFactory,
+    userAgentMiddlewareFactory
+) {
     override val defaultContentType = "application/cbor"
     override val protocol: ShapeId = Rpcv2CborTrait.ID
-    val requestTestBuilder = HttpProtocolUnitTestRequestGenerator.Builder()
-    val responseTestBuilder = HttpProtocolUnitTestResponseGenerator.Builder()
-    val errorTestBuilder = HttpProtocolUnitTestErrorGenerator.Builder()
-    open val protocolTestsToIgnore: Set<String> = setOf()
-    open val protocolTestTagsToIgnore: Set<String> = setOf()
-    private val myRpcCborCustoms = rpcCborCustomizations
-
-    override fun generateProtocolUnitTests(ctx: ProtocolGenerator.GenerationContext): Int {
-        return HttpProtocolTestGenerator(
-            ctx,
-            requestTestBuilder,
-            responseTestBuilder,
-            errorTestBuilder,
-            myRpcCborCustoms,
-            getProtocolHttpBindingResolver(ctx, defaultContentType),
-            protocolTestsToIgnore,
-            protocolTestTagsToIgnore,
-        ).generateProtocolTests() + renderEndpointsTests(ctx) // does not render endpoint tests
-    }
-
-    override val httpProtocolClientGeneratorFactory: HttpProtocolClientGeneratorFactory = SmithyHttpProtocolClientGeneratorFactory()
-
     override val shouldRenderEncodableConformance = true
 
     override fun getProtocolHttpBindingResolver(
@@ -69,14 +38,7 @@ class RpcV2CborProtocolGenerator(
     ): HttpBindingResolver = RPCV2CBORHttpBindingResolver(ctx, defaultContentType)
 
     override fun addProtocolSpecificMiddleware(ctx: ProtocolGenerator.GenerationContext, operation: OperationShape) {
-        val operationEndpointResolverMiddleware = (
-            operationEndpointResolverMiddlewareFactory ?: { _, endpointMiddlewareSymbol -> OperationEndpointResolverMiddleware(ctx, endpointMiddlewareSymbol) }
-            )(ctx, myRpcCborCustoms.endpointMiddlewareSymbol)
-
-        operationMiddleware.appendMiddleware(
-            operation,
-            operationEndpointResolverMiddleware
-        )
+        super.addProtocolSpecificMiddleware(ctx, operation)
 
         operationMiddleware.removeMiddleware(operation, "OperationInputBodyMiddleware")
         operationMiddleware.appendMiddleware(operation, OperationInputBodyMiddleware(ctx.model, ctx.symbolProvider, true))
@@ -118,31 +80,6 @@ class RpcV2CborProtocolGenerator(
         if (!hasEventStreamRequest) {
             operationMiddleware.appendMiddleware(operation, ContentLengthMiddleware(ctx.model, true, false, false))
         }
-    }
-
-    override fun addUserAgentMiddleware(ctx: ProtocolGenerator.GenerationContext, operation: OperationShape) {
-        userAgentMiddlewareFactory?.let { factory ->
-            val userAgentMiddleware = factory(ctx)
-            operationMiddleware.appendMiddleware(operation, userAgentMiddleware)
-        }
-    }
-
-    fun renderEndpointsTests(ctx: ProtocolGenerator.GenerationContext): Int {
-        val ruleSetNode = ctx.service.getTrait<EndpointRuleSetTrait>()?.ruleSet
-        val ruleSet = if (ruleSetNode != null) EndpointRuleSet.fromNode(ruleSetNode) else null
-        var testCount = 0
-
-        ctx.service.getTrait<EndpointTestsTrait>()?.let { testsTrait ->
-            if (testsTrait.testCases?.isEmpty() == true) {
-                return 0
-            }
-
-            ctx.delegator.useFileWriter("Tests/${ctx.settings.testModuleName}/EndpointResolverTest.swift") { swiftWriter ->
-                testCount = + EndpointTestGenerator(testsTrait, ruleSet, ctx).render(swiftWriter)
-            }
-        }
-
-        return testCount
     }
 
     /**
