+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/protocols/rpcv2cbor/SmithyHTTPBindingProtocolGenerator.kt
@@ -1,0 +1,88 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+package software.amazon.smithy.swift.codegen.protocols.rpcv2cbor
+
+import software.amazon.smithy.codegen.core.Symbol
+import software.amazon.smithy.rulesengine.language.EndpointRuleSet
+import software.amazon.smithy.rulesengine.traits.EndpointRuleSetTrait
+import software.amazon.smithy.rulesengine.traits.EndpointTestsTrait
+import software.amazon.smithy.swift.codegen.EndpointTestGenerator
+import software.amazon.smithy.model.shapes.OperationShape
+import software.amazon.smithy.swift.codegen.integration.HTTPBindingProtocolGenerator
+import software.amazon.smithy.swift.codegen.integration.HTTPProtocolCustomizable
+import software.amazon.smithy.swift.codegen.integration.HttpProtocolTestGenerator
+import software.amazon.smithy.swift.codegen.integration.HttpProtocolUnitTestErrorGenerator
+import software.amazon.smithy.swift.codegen.integration.HttpProtocolUnitTestRequestGenerator
+import software.amazon.smithy.swift.codegen.integration.HttpProtocolUnitTestResponseGenerator
+import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
+import software.amazon.smithy.swift.codegen.integration.SmithyHttpProtocolClientGeneratorFactory
+import software.amazon.smithy.swift.codegen.integration.middlewares.OperationEndpointResolverMiddleware
+import software.amazon.smithy.swift.codegen.middleware.MiddlewareRenderable
+import software.amazon.smithy.swift.codegen.model.getTrait
+import software.amazon.smithy.swift.codegen.testModuleName
+
+abstract class SmithyHTTPBindingProtocolGenerator(
+    customizations: HTTPProtocolCustomizable,
+    private val operationEndpointResolverMiddlewareFactory: ((ProtocolGenerator.GenerationContext, Symbol) -> MiddlewareRenderable)? = null,
+    private val userAgentMiddlewareFactory: ((ProtocolGenerator.GenerationContext) -> MiddlewareRenderable)? = null
+) : HTTPBindingProtocolGenerator(customizations) {
+    override val httpProtocolClientGeneratorFactory = SmithyHttpProtocolClientGeneratorFactory()
+
+    val requestTestBuilder = HttpProtocolUnitTestRequestGenerator.Builder()
+    val responseTestBuilder = HttpProtocolUnitTestResponseGenerator.Builder()
+    val errorTestBuilder = HttpProtocolUnitTestErrorGenerator.Builder()
+    open val protocolTestsToIgnore: Set<String> = setOf()
+    open val protocolTestTagsToIgnore: Set<String> = setOf()
+
+    override val shouldRenderEncodableConformance = false
+    override fun generateProtocolUnitTests(ctx: ProtocolGenerator.GenerationContext): Int {
+        return HttpProtocolTestGenerator(
+            ctx,
+            requestTestBuilder,
+            responseTestBuilder,
+            errorTestBuilder,
+            customizations,
+            getProtocolHttpBindingResolver(ctx, defaultContentType),
+            protocolTestsToIgnore,
+            protocolTestTagsToIgnore,
+        ).generateProtocolTests() + renderEndpointsTests(ctx)
+    }
+
+    fun renderEndpointsTests(ctx: ProtocolGenerator.GenerationContext): Int {
+        val ruleSetNode = ctx.service.getTrait<EndpointRuleSetTrait>()?.ruleSet
+        val ruleSet = if (ruleSetNode != null) EndpointRuleSet.fromNode(ruleSetNode) else null
+        var testCount = 0
+
+        ctx.service.getTrait<EndpointTestsTrait>()?.let { testsTrait ->
+            if (testsTrait.testCases?.isEmpty() == true) {
+                return 0
+            }
+
+            ctx.delegator.useFileWriter("Tests/${ctx.settings.testModuleName}/EndpointResolverTest.swift") { swiftWriter ->
+                testCount = + EndpointTestGenerator(testsTrait, ruleSet, ctx).render(swiftWriter)
+            }
+        }
+
+        return testCount
+    }
+
+    override fun addProtocolSpecificMiddleware(ctx: ProtocolGenerator.GenerationContext, operation: OperationShape) {
+        val operationEndpointResolverMiddleware = (
+            operationEndpointResolverMiddlewareFactory ?: { _, endpointMiddlewareSymbol -> OperationEndpointResolverMiddleware(ctx, endpointMiddlewareSymbol) }
+            )(ctx, customizations.endpointMiddlewareSymbol)
+
+        operationMiddleware.appendMiddleware(
+            operation,
+            operationEndpointResolverMiddleware
+        )
+    }
+
+    override fun addUserAgentMiddleware(ctx: ProtocolGenerator.GenerationContext, operation: OperationShape) {
+        userAgentMiddlewareFactory?.let { factory ->
+            val userAgentMiddleware = factory(ctx)
+            operationMiddleware.appendMiddleware(operation, userAgentMiddleware)
+        }
+    }
+}
